Compiles with GCC
Set COMPILER in your Makelocal to GCC if you want to not compile with
Ivy.  It will compile and run with Ivy or gcc, though there were a few
places where they are mutually unhappy, or we couldn't figure out nice
ways to do the annotations.  For instance, the struct work isn't
supposed to be TP(env_t*) all the time.  Also, cil destroys our attempts
at unbounded arrays.  Look for pragma nodeputy in a few places.

Also changed types so we can eventually have our own, um, i guess we
still need to have our own types.h for the kernel, but we cleaned it up a
bit is the point, right? (Kevin, verbatim)

Some work in progress for mmap and other stuff, but I never did any of
that.  Just ignore it for now.

+++ kern/src/kfs.c
@@ -64,7 +64,7 @@ ssize_t kfs_lookup_path(char* path)
  * This should take a real inode or something to point to the real location,
  * and env_create shouldn't assume everything is contiguous
  */
-struct proc *kfs_proc_create(size_t kfs_inode)
+struct proc *kfs_proc_create(int kfs_inode)
 {
 	if (kfs_inode < 0 || kfs_inode >= MAX_KFS_FILES)
 		panic("Invalid kfs_inode.  Check you error codes!");
